consolidated code into map.jsx... markers are clickable and updating state correctly
